+++ testutil/fund.go
@@ -19,12 +19,12 @@ package testutil
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
+	"github.com/evmos/evmos/v11/utils"
 	inflationtypes "github.com/evmos/evmos/v11/x/inflation/types"
 )
 
 // FundAccount is a utility function that funds an account by minting and
-// sending the coins to the address. This should be used for testing purposes
-// only!
+// sending the coins to the address.
 func FundAccount(ctx sdk.Context, bankKeeper bankkeeper.Keeper, addr sdk.AccAddress, amounts sdk.Coins) error {
 	if err := bankKeeper.MintCoins(ctx, inflationtypes.ModuleName, amounts); err != nil {
 		return err
@@ -33,9 +33,17 @@ func FundAccount(ctx sdk.Context, bankKeeper bankkeeper.Keeper, addr sdk.AccAddr
 	return bankKeeper.SendCoinsFromModuleToAccount(ctx, inflationtypes.ModuleName, addr, amounts)
 }
 
+// FundAccountWithBaseDenom is a utility function that uses the FundAccount function
+// to fund an account with the default Evmos denomination.
+func FundAccountWithBaseDenom(ctx sdk.Context, bankKeeper bankkeeper.Keeper, addr sdk.AccAddress, amount int64) error {
+	coins := sdk.NewCoins(
+		sdk.NewCoin(utils.BaseDenom, sdk.NewInt(amount)),
+	)
+	return FundAccount(ctx, bankKeeper, addr, coins)
+}
+
 // FundModuleAccount is a utility function that funds a module account by
-// minting and sending the coins to the address. This should be used for testing
-// purposes only!
+// minting and sending the coins to the address.
 func FundModuleAccount(ctx sdk.Context, bankKeeper bankkeeper.Keeper, recipientMod string, amounts sdk.Coins) error {
 	if err := bankKeeper.MintCoins(ctx, inflationtypes.ModuleName, amounts); err != nil {
 		return err